qcacmn: [HIF] Add support for QCA9379 on x86 platform
Add support for QCA9379 chip on x86 platform for
USB and SDIO attach.

CRs-Fixed: 2181012

+++ hif/src/sdio/hif_sdio_send.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013-2017 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2013-2018 The Linux Foundation. All rights reserved.
  *
  * Previously licensed under the ISC license by Qualcomm Atheros, Inc.
  *
@@ -156,7 +156,7 @@ QDF_STATUS hif_dev_send_buffer(struct hif_sdio_device *pdev,
 		} else {
 			AR_DEBUG_PRINTF(ATH_DEBUG_ERR,
 				("Allocate send context fail %d\n",
-				sizeof(struct hif_sendContext) +
+				(int) sizeof(struct hif_sendContext) +
 				padded_length));
 			return QDF_STATUS_E_NOMEM;
 		}
